# Add Katana to performance comparison