--- a/app/build.gradle
+++ b/app/build.gradle
@@ -36,4 +36,7 @@ dependencies {
     // Dagger
     implementation 'com.google.dagger:dagger:2.16'
     kapt 'com.google.dagger:dagger-compiler:2.16'
+
+    // Katana
+    implementation 'com.github.rewe-digital-incubator.katana:katana-core:1.0.1'
 }