… app/src/main/java/com/sloydev/dependencyinjectionperformance/katana/KatanaJavaModule.kt
… app/src/main/java/com/sloydev/dependencyinjectionperformance/katana/KatanaKotlinModule.kt