Please update `README.md` with the test results of your devices.
Here are my results:

```
Google Pixel (Android 9.0)

=========|=====================
Device:  | google sailfish v9
---------|--------------------
Test:    | Koin + Kotlin
Min-Max: | 4,45-48,46 ms
Average: | 6,75 ms
---------|--------------------
Test:    | Koin + Java
Min-Max: | 5,57-12,63 ms
Average: | 6,56 ms
---------|--------------------
Test:    | Kodein + Kotlin
Min-Max: | 0,23-6,97 ms
Average: | 0,65 ms
---------|--------------------
Test:    | Kodein + Java
Min-Max: | 0,19-2,45 ms
Average: | 0,36 ms
---------|--------------------
Test:    | Dagger2 + Kotlin
Min-Max: | 0,03-2,64 ms
Average: | 0,12 ms
---------|--------------------
Test:    | Dagger2 + Java
Min-Max: | 0,03-2,50 ms
Average: | 0,11 ms
---------|--------------------
Test:    | Katana + Kotlin
Min-Max: | 0,06-1,92 ms
Average: | 0,31 ms
---------|--------------------
Test:    | Katana + Java
Min-Max: | 0,05-0,44 ms
Average: | 0,12 ms
=========|=====================


Google Pixel 2 XL (Android 9.0)
 
=========|=====================
Device:  | google taimen v9
---------|--------------------
Test:    | Koin + Kotlin
Min-Max: | 3,92-24,05 ms
Average: | 4,81 ms
---------|--------------------
Test:    | Koin + Java
Min-Max: | 4,01-7,07 ms
Average: | 4,57 ms
---------|--------------------
Test:    | Kodein + Kotlin
Min-Max: | 0,08-4,13 ms
Average: | 0,26 ms
---------|--------------------
Test:    | Kodein + Java
Min-Max: | 0,08-1,36 ms
Average: | 0,16 ms
---------|--------------------
Test:    | Dagger2 + Kotlin
Min-Max: | 0,02-1,47 ms
Average: | 0,06 ms
---------|--------------------
Test:    | Dagger2 + Java
Min-Max: | 0,02-1,36 ms
Average: | 0,06 ms
---------|--------------------
Test:    | Katana + Kotlin
Min-Max: | 0,03-1,06 ms
Average: | 0,28 ms
---------|--------------------
Test:    | Katana + Java
Min-Max: | 0,05-0,41 ms
Average: | 0,12 ms
=========|=====================
```