@Sloy 2018-11-28
No! Omg I can't believe I missed your PR. And your issue!! What happened to youtube notifications? 😱😢
I actually talked to someone about how nice would it be to add Katana to the project.

I'll check this out first thing tomorrow morning and update the benchmarks. Thanks a lot for contributing, and sorry again.